Added basic printing functionality
