+++ Main/App.xaml.cs
@@ -10,9 +10,10 @@ namespace DrawingNameComposer
 	/// </summary>
 	public partial class App : Application
 	{
-		protected override async void OnStartup(StartupEventArgs e)
+		protected override void OnStartup(StartupEventArgs e)
 		{
-			var mainViewModel = new MainViewModel(new MetadataService(), new PrintingService(), new PresetsService(new AppPathsProvider()));
+			var printingSettingsService = new PrintingSettingsService();
+			var mainViewModel = new MainViewModel(new MetadataService(), printingSettingsService, new PresetsService(new AppPathsProvider()), new PrintingService(printingSettingsService));
 			mainViewModel.Initialize();
 			var mainWindow = new MainWindow(mainViewModel);
 			mainWindow.Show();
