+++ Main/DrawingNameComposer.csproj
@@ -22,6 +22,7 @@
 		<PackageReference Include="CommunityToolkit.Mvvm" Version="8.3.2" />
 		<PackageReference Include="gong-wpf-dragdrop" Version="3.2.1" />
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+		<PackageReference Include="PDFsharp" Version="6.1.1" />
 		<PackageReference Include="Tekla.Structures" Version="2024.0.2" />
 		<PackageReference Include="Tekla.Structures.Drawing" Version="2024.0.2" />
 		<PackageReference Include="Tekla.Structures.Model" Version="2024.0.2" />
