--- Main/Services/PrintingService.cs
@@ -1,9 +1,0 @@
-﻿namespace DrawingNameComposer.Services;
-
-public class PrintingService : IPrintingService
-{
-	public IEnumerable<string> Get()
-	{
-		return [];
-	}
-}
